+++ R/batch-chat.R
@@ -2,11 +2,13 @@
 #'
 #' @description
 #' `batch_chat()` and `batch_chat_structured()` currently only work with
-#' [chat_openai()], [chat_anthropic()], and [chat_google_gemini()]. They use
-#' the [OpenAI](https://platform.openai.com/docs/guides/batch),
+#' [chat_openai()], [chat_anthropic()], [chat_google_gemini()], and
+#' [chat_groq()]. They use the
+#' [OpenAI](https://platform.openai.com/docs/guides/batch),
 #' [Anthropic](https://docs.claude.com/en/docs/build-with-claude/batch-processing),
-#' and [Google Gemini](https://ai.google.dev/gemini-api/docs/batch-api) batch APIs
-#' which allow you to submit multiple requests simultaneously.
+#' [Google Gemini](https://ai.google.dev/gemini-api/docs/batch-api), and
+#' [Groq](https://console.groq.com/docs/batch) batch APIs which allow you
+#' to submit multiple requests simultaneously.
 #' The results can take up to 24 hours to complete, but in return you pay 50%
 #' less than usual (but note that ellmer doesn't include this discount in
 #' its pricing metadata). If you want to get results back more quickly, or