+++ tests/testthat/test-provider-groq.R
@@ -20,3 +20,54 @@ test_that("supports structured data", {
   }
   test_data_extraction(chat_fun)
 })
+
+test_that("batch chat works", {
+  chat <- chat_groq(
+    system_prompt = "Answer with just the city name",
+    model = "llama-3.1-8b-instant",
+    params = params(temperature = 0, seed = 1014),
+    credentials = function() list(Authorization = "Bearer x")
+  )
+
+  prompts <- list(
+    "What's the capital of Iowa?",
+    "What's the capital of New York?",
+    "What's the capital of California?",
+    "What's the capital of Texas?"
+  )
+
+  out <- batch_chat_text(
+    chat,
+    prompts,
+    path = test_path("batch/state-capitals-groq.json"),
+    ignore_hash = TRUE
+  )
+  expect_equal(out, c("Des Moines", "Albany", "Sacramento", "Austin"))
+})
+
+test_that("batch_chat_structured works", {
+  chat <- chat_groq(
+    system_prompt = "Answer with just the city name",
+    model = "openai/gpt-oss-20b",
+    params = params(temperature = 0, seed = 1014),
+    credentials = function() list(Authorization = "Bearer x")
+  )
+
+  prompts <- list(
+    "What's the capital of Iowa?",
+    "What's the capital of New York?",
+    "What's the capital of California?",
+    "What's the capital of Texas?"
+  )
+
+  out <- batch_chat_structured(
+    chat,
+    prompts,
+    path = test_path("batch/state-capitals-groq-structured.json"),
+    type = type_object(capital = type_string()),
+    ignore_hash = TRUE
+  )
+
+  expect_s3_class(out, "data.frame")
+  expect_equal(out$capital, c("Des Moines", "Albany", "Sacramento", "Austin"))
+})